feat: replace repository table with RepositoriesTable component
